requirements.txt
flask
flask-cors
flask-sqlalchemy
pymysql
cryptography
tensorflow
numpy
Pillow
opencv-python
google-genai
python-dotenv
werkzeug
matplotlib
scikit-learn
setuptools